Cleanup headers, move external libraries on top, internal bottom, removed unused imports
