+++ src/Audio_TD5580A.h
@@ -2,9 +2,8 @@
 #define PT_AUDIO
 
 #include <Arduino.h>
-#include "Settings.h"
 #include <SerialMP3.h>
-
+#include "Settings.h"
 #include "Pins.h"
 
 class Audio {
