+++ src/ESP32Downloader.cpp
@@ -1,9 +1,7 @@
 #ifdef HARDWARE_V3
-#include <WiFi.h>
 #include <HTTPClient.h>
 #include <LittleFS.h>
-#include "config.h"
-
+#include <WiFi.h>
 #include <WiFiClient.h>
 WiFiClient* wifiClient;
